Accepted: didnt check for float, was mentioned in question, read question carefully

+++ Solved Problems/Easy/2465_Number_of_Distinct_Averages.cpp
@@ -1,11 +1,11 @@
 class Solution {
     public:
         int distinctAverages(vector<int>& nums) {
-            set<int> s;
+            set<float> s;
             sort(nums.begin(), nums.end());
             int i = 0, j = nums.size() - 1;
             while(i < j) {
-                s.insert((nums[i] + nums[j]) / 2);
+                s.insert((nums[i] + nums[j]) / 2.0);
                 i++;
                 j--;
             }
